combine token refresher service and auth service
