+++ prediction-polls/android/app/src/main/java/com/bounswe/predictionpolls/data/remote/repositories/AuthRepository.kt
@@ -28,4 +28,10 @@ class AuthRepository @Inject constructor(
             tokenManager.refreshToken = null
         }
     }
+
+    suspend fun refreshAccessToken(): String? {
+        val newToken = authService.refreshAccessToken().body()?.accessToken
+        tokenManager.accessToken = newToken
+        return newToken
+    }
 }
