+++ prediction-polls/android/app/src/main/java/com/bounswe/predictionpolls/data/remote/services/AuthService.kt
@@ -1,6 +1,7 @@
 package com.bounswe.predictionpolls.data.remote.services
 
 import com.bounswe.predictionpolls.data.remote.response.LoginResponse
+import com.bounswe.predictionpolls.data.remote.response.RefreshAccessTokenResponse
 import retrofit2.Response
 import retrofit2.http.POST
 
@@ -13,4 +14,7 @@ interface AuthService {
 
     @POST("/logout")
     suspend fun logout(): Response<Void>
+
+    @POST("/access-token")
+    suspend fun refreshAccessToken(): Response<RefreshAccessTokenResponse>
 }
