+++ prediction-polls/android/app/src/main/java/com/bounswe/predictionpolls/di/RepositoryModule.kt
@@ -2,9 +2,7 @@ package com.bounswe.predictionpolls.di
 
 import com.bounswe.predictionpolls.data.remote.TokenManager
 import com.bounswe.predictionpolls.data.remote.repositories.AuthRepository
-import com.bounswe.predictionpolls.data.remote.repositories.TokenRefresherRepository
 import com.bounswe.predictionpolls.data.remote.services.AuthService
-import com.bounswe.predictionpolls.data.remote.services.TokenRefresherService
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -21,14 +19,5 @@ object RepositoryModule {
         tokenManager: TokenManager
     ): AuthRepository {
         return AuthRepository(authService, tokenManager)
-    }
-
-    @Provides
-    @Singleton
-    fun provideTokenRefresherRepository(
-        tokenRefresherService: TokenRefresherService,
-        tokenManager: TokenManager
-    ): TokenRefresherRepository {
-        return TokenRefresherRepository(tokenRefresherService, tokenManager)
     }
 }
