+++ prediction-polls/android/app/src/main/java/com/bounswe/predictionpolls/di/ServiceModule.kt
@@ -1,7 +1,6 @@
 package com.bounswe.predictionpolls.di
 
 import com.bounswe.predictionpolls.data.remote.services.AuthService
-import com.bounswe.predictionpolls.data.remote.services.TokenRefresherService
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -15,16 +14,8 @@ object ServiceModule {
     @Provides
     @Singleton
     fun provideAuthService(
-        @AuthenticatedRetrofit retrofit: Retrofit
+        @UnauthenticatedRetrofit retrofit: Retrofit
     ): AuthService {
         return retrofit.create(AuthService::class.java)
-    }
-
-    @Provides
-    @Singleton
-    fun provideTokenRefresherService(
-        @TokenRefresherRetrofit retrofit: Retrofit
-    ): TokenRefresherService {
-        return retrofit.create(TokenRefresherService::class.java)
     }
 }
